+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/), and this
 
 ### Added
 
+* Supports for 'X-Forwarded-Host/Forwarded' headers when behind a proxy
 * Allow 'X-Qgis-Project' header for passing project path
 
 ### Fixed